requirements.txt
eventlet~=0.36.1
python-socketio